feat: 유저 수를 나타내는 숫자를 채팅 헤더에서 사이드바 유저아이콘으로 이동
