+++ next/src/components/video-chat/ChatSection/Container/index.tsx
@@ -1,14 +1,10 @@
-import React, { useEffect, useState, useContext } from 'react';
-import { ChatEvent } from '@lib/enum';
-import useChatClient from '@hooks/chat/useChatClient';
+import React, { useState, useContext } from 'react';
 import { SideSectionShowContext } from '@components/video-chat/Container';
 
 import useDropzone from './useDropzone';
 import * as S from './style';
 
 const Container: React.FC = ({ children }) => {
-  const socketClient = useChatClient();
-  const [memberCount, setMemberCount] = useState<number>(0);
   const [sideSectionShowState, toggleSideSection] = useContext(
     SideSectionShowContext,
   );
@@ -18,16 +14,6 @@ const Container: React.FC = ({ children }) => {
   const handleChatCloseButtonClick = () => {
     toggleSideSection('chat', { justOff: true });
   };
-
-  useEffect(() => {
-    socketClient.on(ChatEvent.GET_MEMBER_COUNT, (count: number) => {
-      setMemberCount(count);
-    });
-
-    return () => {
-      socketClient.off(ChatEvent.GET_MEMBER_COUNT);
-    };
-  }, [socketClient]);
 
   return (
     <div {...getRootProps()}>
@@ -39,7 +25,6 @@ const Container: React.FC = ({ children }) => {
         <S.Header>
           <S.ChatCloseButton onClick={handleChatCloseButtonClick} />
           <S.ChatTitle>채팅</S.ChatTitle>
-          <S.ChatMemberCount>{memberCount}</S.ChatMemberCount>
         </S.Header>
 
         {children}
