+++ next/src/components/video-chat/Sidebar/index.tsx
@@ -1,15 +1,20 @@
-import { useContext } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import _ from 'lodash';
+
+import useChatClient from '@hooks/chat/useChatClient';
+import { ChatEvent } from '@lib/enum';
 import Profile from '@components/common/Profile';
 import { SideSectionShowContext } from '@components/video-chat/Container';
-
 import SVGButton from '@components/common/SVGButton';
+
 import * as S from './style';
 
 const Sidebar = () => {
+  const socketClient = useChatClient();
   const [sideSectionShowState, toggleSideSection] = useContext(
     SideSectionShowContext,
   );
+  const [memberCount, setMemberCount] = useState<number>(0);
 
   const handleChatIconClick = () => {
     toggleSideSection('chat');
@@ -22,6 +27,16 @@ const Sidebar = () => {
   const handleUserIconClick = () => {
     toggleSideSection('userList');
   };
+
+  useEffect(() => {
+    socketClient.on(ChatEvent.GET_MEMBER_COUNT, (count: number) => {
+      setMemberCount(count);
+    });
+
+    return () => {
+      socketClient.off(ChatEvent.GET_MEMBER_COUNT);
+    };
+  }, [socketClient]);
 
   return (
     <S.Sidebar>
@@ -45,16 +60,19 @@ const Sidebar = () => {
           }}
           onClick={handleChatIconClick}
         />
-        <SVGButton
-          SvgComponent={S.UserIcon}
-          buttonProps={{
-            title: sideSectionShowState.userList
-              ? '사용자 목록 닫기'
-              : '사용자 목록 열기',
-            'aria-label': 'Toggle user list',
-          }}
-          onClick={handleUserIconClick}
-        />
+        <S.RelativeArea>
+          <SVGButton
+            SvgComponent={S.UserIcon}
+            buttonProps={{
+              title: sideSectionShowState.userList
+                ? '사용자 목록창 닫기'
+                : '사용자 목록창 열기',
+              'aria-label': 'Toggle user list',
+            }}
+            onClick={handleUserIconClick}
+          />
+          <S.MemberCount>{memberCount}</S.MemberCount>
+        </S.RelativeArea>
       </S.BottomSection>
     </S.Sidebar>
   );
